Add dependencies for the contact form
This requires a number directly, and a few indirectly. The form's
generated based on Zend Form annotations, so needs Doctrine Annotations
and Zend Form. It's also making use of validation and input filters, so
those libraries are also required. Faker's being added to make it easy
to generate test data.